requirements.txt
fastapi~=0.116.1
uvicorn~=0.35.0
SQLAlchemy~=2.0.43
pydantic~=2.11.7
passlib~=1.7.4
bcrypt
pyjwt~=2.10.1
python-jose~=3.5.0